Split popover desktop and mobile classes
