+++ src/components/ui/toolbox.ts
@@ -5,10 +5,12 @@ import BlockTool from '../tools/block';
 import ToolsCollection from '../tools/collection';
 import { API, BlockToolData, ToolboxConfigEntry, PopoverItem, BlockAPI } from '../../../types';
 import EventsDispatcher from '../utils/events';
-import Popover from '../utils/popover';
 import I18n from '../i18n';
 import { I18nInternalNS } from '../i18n/namespace-internal';
 import { PopoverEvent } from '../utils/popover/popover.typings';
+import Listeners from '../utils/listeners';
+import Dom from '../dom';
+import { PopoverDesktop, PopoverMobile } from '../utils/popover';
 
 /**
  * @todo the first Tab on the Block — focus Plus Button, the second — focus Block Tunes Toggler, the third — focus next Block
@@ -77,6 +79,11 @@ export default class Toolbox extends EventsDispatcher<ToolboxEventMap> {
   public opened = false;
 
   /**
+   * Listeners util instance
+   */
+  protected listeners: Listeners = new Listeners();
+
+  /**
    * Editor API
    */
   private api: API;
@@ -84,7 +91,7 @@ export default class Toolbox extends EventsDispatcher<ToolboxEventMap> {
   /**
    * Popover instance. There is a util for vertical lists.
    */
-  private popover: Popover | undefined;
+  private popover: PopoverDesktop | PopoverMobile | undefined;
 
   /**
    * List of Tools available. Some of them will be shown in the Toolbox
@@ -129,36 +136,13 @@ export default class Toolbox extends EventsDispatcher<ToolboxEventMap> {
     this.api = api;
     this.tools = tools;
     this.i18nLabels = i18nLabels;
+    this.make();
   }
 
   /**
-   * Makes the Toolbox
+   * Returns root block settings element
    */
-  public make(): Element {
-    this.popover = new Popover({
-      scopeElement: this.api.ui.nodes.redactor,
-      searchable: true,
-      messages: {
-        nothingFound: this.i18nLabels.nothingFound,
-        search: this.i18nLabels.filter,
-      },
-      items: this.toolboxItemsToBeDisplayed,
-    });
-
-    this.popover.on(PopoverEvent.Close, this.onPopoverClose);
-
-    /**
-     * Enable tools shortcuts
-     */
-    this.enableShortcuts();
-
-    this.nodes.toolbox = this.popover.getElement();
-    this.nodes.toolbox.classList.add(Toolbox.CSS.toolbox);
-
-    if (import.meta.env.MODE === 'test') {
-      this.nodes.toolbox.setAttribute('data-cy', 'toolbox');
-    }
-
+  public getElement(): HTMLElement {
     return this.nodes.toolbox;
   }
 
@@ -166,7 +150,7 @@ export default class Toolbox extends EventsDispatcher<ToolboxEventMap> {
    * Returns true if the Toolbox has the Flipper activated and the Flipper has selected button
    */
   public hasFocus(): boolean | undefined {
-    return this.popover?.hasFocus();
+    return 'hasFocus' in this.popover ? this.popover?.hasFocus() : undefined;
   }
 
   /**
@@ -182,6 +166,7 @@ export default class Toolbox extends EventsDispatcher<ToolboxEventMap> {
 
     this.removeAllShortcuts();
     this.popover?.off(PopoverEvent.Close, this.onPopoverClose);
+    this.listeners.destroy();
   }
 
   /**
@@ -225,6 +210,64 @@ export default class Toolbox extends EventsDispatcher<ToolboxEventMap> {
     } else {
       this.close();
     }
+  }
+
+  /**
+   * Destroys existing popover instance and contructs the new one.
+   * Is needed to
+   */
+  public forceRemountPopover = (): void  => {
+    this.destroyPopover();
+    this.initPopover();
+  };
+
+  /**
+   * Makes the Toolbox
+   */
+  private make(): void {
+    /**
+     * Enable tools shortcuts
+     */
+    this.enableShortcuts();
+
+    this.nodes.toolbox = Dom.make('div', Toolbox.CSS.toolbox);
+
+    this.initPopover();
+
+    if (import.meta.env.MODE === 'test') {
+      this.nodes.toolbox.setAttribute('data-cy', 'toolbox');
+    }
+  }
+
+  /**
+   * Creates toolbox popover and appends it inside wrapper element
+   */
+  private initPopover(): void {
+    const Popover = _.isMobileScreen() ? PopoverMobile : PopoverDesktop;
+
+    this.popover = new Popover({
+      scopeElement: this.api.ui.nodes.redactor,
+      searchable: true,
+      messages: {
+        nothingFound: this.i18nLabels.nothingFound,
+        search: this.i18nLabels.filter,
+      },
+      items: this.toolboxItemsToBeDisplayed,
+    });
+
+    this.popover.on(PopoverEvent.Close, this.onPopoverClose);
+    this.nodes.toolbox.append(this.popover.getElement());
+  }
+
+  /**
+   * Destroys popover instance and removes it from DOM
+   */
+  private destroyPopover(): void {
+    this.popover.hide();
+    this.popover.off(PopoverEvent.Close, this.onPopoverClose);
+    this.popover.destroy();
+    this.popover = null;
+    this.nodes.toolbox.innerHTML = '';
   }
 
   /**
